Fix formating of title and name.
There were problems when outputing to the gitbash console. This led to
the encode() being added. But that then caused the "b" prefix (bytes).
So the change was reverted. But I am not sure if the old problem is now
present again (the bugs that triggered the problem are now closed).